+++ settings/lib/model/developer_settings_model.dart
@@ -30,4 +30,4 @@ class DeveloperSettingsModel {
   }
 }
 
-enum Feature { logging }
+enum Feature { placeholder }