+++ station/lib/repository/marker_repository.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:core/log/log.dart';
 import 'package:multiple_result/multiple_result.dart';
 import 'package:navigation/coordinate_model.dart';
 import 'package:station/model/config_model.dart';
@@ -48,7 +49,9 @@ class TanksteWebMarkerRepository extends MarkerRepository {
       Result<ConfigModel, Exception> configResult =
           await _configRepository.get().first;
       if (configResult.isError()) {
-        return Result.error(configResult.tryGetError()!);
+        Exception error = configResult.tryGetError()!;
+        Log.exception(error);
+        return Result.error(error);
       }
       ConfigModel config = configResult.tryGetSuccess()!;
 
@@ -70,9 +73,12 @@ class TanksteWebMarkerRepository extends MarkerRepository {
 
         return Result.success(markers);
       } else {
-        return Result.error(Exception("API Error!\n\n${response.body}"));
+        Exception error = Exception("API Error!\n\n${response.body}");
+        Log.exception(error);
+        return Result.error(error);
       }
     } on Exception catch (e) {
+      Log.exception(e);
       return Result.error(e);
     }
   }