+++ station/lib/repository/open_time_repository.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'package:core/log/log.dart';
 import 'package:http/http.dart' as http;
 import 'package:multiple_result/multiple_result.dart';
 import 'package:station/model/config_model.dart';
@@ -35,7 +36,9 @@ class TanksteWebOpenTimeRepository extends OpenTimeRepository {
       Result<ConfigModel, Exception> configResult =
           await _configRepository.get().first;
       if (configResult.isError()) {
-        return Result.error(configResult.tryGetError()!);
+        Exception error = configResult.tryGetError()!;
+        Log.exception(error);
+        return Result.error(error);
       }
       ConfigModel config = configResult.tryGetSuccess()!;
 
@@ -54,9 +57,12 @@ class TanksteWebOpenTimeRepository extends OpenTimeRepository {
 
         return Result.success(openTimes);
       } else {
-        return Result.error(Exception("API Error!\n\n${response.body}"));
+        Exception error = Exception("API Error!\n\n${response.body}");
+        Log.exception(error);
+        return Result.error(error);
       }
     } on Exception catch (e) {
+      Log.exception(e);
       return Result.error(e);
     }
   }